Lager en uniform respons ut av service kall.
